Revert "feat(web): move emails notifications to one pane and enable opening settings via url (#1614)"
This reverts commit 1fe833bc24a3a426a7230381704d418624cc2b82.
